isolated-sstate-dir: only enable if ISOLATED_SSTATE_DIR != SSTATE_DIR

+++ classes/isolated-sstate-dir.bbclass
@@ -6,7 +6,8 @@ ISOLATED_SSTATE_DIR ?= "${TMPDIR}/sstate-cache"
 ISOLATED_SSTATE_PATHSPEC = "${@SSTATE_PATHSPEC.replace(SSTATE_DIR, ISOLATED_SSTATE_DIR)}"
 
 sstate_write_isolated () {
-    if [ -n "${ISOLATED_SSTATE_DIR}" ]; then
+    if [ -n "${ISOLATED_SSTATE_DIR}" ] && \
+       [ "${ISOLATED_SSTATE_DIR}" != "${SSTATE_DIR}" ]; then
         isolated_dest=$(echo ${SSTATE_PKG} | sed 's,^${SSTATE_DIR}/,${ISOLATED_SSTATE_DIR}/,')
         mkdir -p $(dirname $isolated_dest)
         rm -f $isolated_dest $isolated_dest.siginfo
@@ -31,7 +32,7 @@ SSTATEPREINSTFUNCS += "sstate_write_isolated_preinst"
 
 def cleansstate_isolated(d):
     isolated = d.getVar('ISOLATED_SSTATE_DIR', True)
-    if isolated:
+    if isolated and isolated != d.getVar('SSTATE_DIR', True):
         for task in (d.getVar('SSTATETASKS', True) or "").split():
                 ss = sstate_state_fromvars(d, task[3:])
                 sstatepkgfile = d.getVar('ISOLATED_SSTATE_PATHSPEC', True) + "*_" + ss['name'] + ".tgz*"
